@@ -19,7 +19,7 @@ use FOS\OAuthServerBundle\Model\ClientInterface;
 use FOS\OAuthServerBundle\Model\ClientManagerInterface;
 use OAuth2\OAuth2;
 use OAuth2\OAuth2ServerException;
-use Symfony\Bundle\FrameworkBundle\Templating\EngineInterface;
+use RuntimeException;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\Form\Form;
 use Symfony\Component\HttpFoundation\Request;
@@ -31,6 +31,7 @@ use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 use Symfony\Component\Security\Core\User\UserInterface;
+use Twig\Environment;
 
 /**
  * Controller handling basic authorization.
@@ -65,9 +66,9 @@ class AuthorizeController
     private $oAuth2Server;
 
     /**
-     * @var EngineInterface
+     * @var Environment
      */
-    private $templating;
+    private $twig;
 
     /**
      * @var RequestStack
@@ -90,11 +91,6 @@ class AuthorizeController
     private $clientManager;
 
     /**
-     * @var string
-     */
-    private $templateEngineType;
-
-    /**
      * @var EventDispatcherInterface
      */
     private $eventDispatcher;
@@ -105,47 +101,32 @@ class AuthorizeController
      *
      * @todo This controller could be refactored to not rely on so many dependencies
      *
-     * @param RequestStack             $requestStack
-     * @param Form                     $authorizeForm
-     * @param AuthorizeFormHandler     $authorizeFormHandler
-     * @param OAuth2                   $oAuth2Server
-     * @param EngineInterface          $templating
-     * @param TokenStorageInterface    $tokenStorage
-     * @param UrlGeneratorInterface    $router
-     * @param ClientManagerInterface   $clientManager
-     * @param EventDispatcherInterface $eventDispatcher
-     * @param SessionInterface         $session
-     * @param string                   $templateEngineType
+     * @param SessionInterface $session
      */
     public function __construct(
         RequestStack $requestStack,
         Form $authorizeForm,
         AuthorizeFormHandler $authorizeFormHandler,
         OAuth2 $oAuth2Server,
-        EngineInterface $templating,
+        Environment $twig,
         TokenStorageInterface $tokenStorage,
         UrlGeneratorInterface $router,
         ClientManagerInterface $clientManager,
         EventDispatcherInterface $eventDispatcher,
-        SessionInterface $session = null,
-        $templateEngineType = 'twig'
+        SessionInterface $session = null
     ) {
         $this->requestStack = $requestStack;
         $this->session = $session;
         $this->authorizeForm = $authorizeForm;
         $this->authorizeFormHandler = $authorizeFormHandler;
         $this->oAuth2Server = $oAuth2Server;
-        $this->templating = $templating;
+        $this->twig = $twig;
         $this->tokenStorage = $tokenStorage;
         $this->router = $router;
         $this->clientManager = $clientManager;
-        $this->templateEngineType = $templateEngineType;
         $this->eventDispatcher = $eventDispatcher;
     }
 
-    /**
-     * Authorize.
-     */
     public function authorizeAction(Request $request)
     {
         $user = $this->tokenStorage->getToken()->getUser();
@@ -164,8 +145,11 @@ class AuthorizeController
 
         /** @var OAuthEvent $event */
         $event = $this->eventDispatcher->dispatch(
-            OAuthEvent::PRE_AUTHORIZATION_PROCESS,
-            new OAuthEvent($user, $this->getClient())
+            new OAuthEvent(
+                $user,
+                $this->getClient()
+            ),
+            OAuthEvent::PRE_AUTHORIZATION_PROCESS
         );
 
         if ($event->isAuthorizedClient()) {
@@ -183,26 +167,25 @@ class AuthorizeController
             'client' => $this->getClient(),
         ];
 
-        return $this->renderAuthorize($data, $this->templating, $this->templateEngineType);
+        return new Response(
+            $this->twig->render('@FOSOAuthServer/Authorize/authorize.html.twig', $data),
+            Response::HTTP_OK
+        );
     }
 
-    /**
-     * @param UserInterface        $user
-     * @param AuthorizeFormHandler $formHandler
-     * @param Request              $request
-     *
-     * @return Response
-     */
-    protected function processSuccess(UserInterface $user, AuthorizeFormHandler $formHandler, Request $request)
-    {
+    protected function processSuccess(
+        UserInterface $user,
+        AuthorizeFormHandler $formHandler,
+        Request $request
+    ): ?Response {
         if ($this->session && true === $this->session->get('_fos_oauth_server.ensure_logout')) {
             $this->tokenStorage->setToken(null);
             $this->session->invalidate();
         }
 
         $this->eventDispatcher->dispatch(
-            OAuthEvent::POST_AUTHORIZATION_PROCESS,
-            new OAuthEvent($user, $this->getClient(), $formHandler->isAccepted())
+            new OAuthEvent($user, $this->getClient(), $formHandler->isAccepted()),
+            OAuthEvent::POST_AUTHORIZATION_PROCESS
         );
 
         $formName = $this->authorizeForm->getName();
@@ -221,20 +204,13 @@ class AuthorizeController
 
     /**
      * Generate the redirection url when the authorize is completed.
-     *
-     * @param UserInterface $user
-     *
-     * @return string
      */
-    protected function getRedirectionUrl(UserInterface $user)
+    protected function getRedirectionUrl(UserInterface $user): string
     {
         return $this->router->generate('fos_oauth_server_profile_show');
     }
 
-    /**
-     * @return ClientInterface
-     */
-    protected function getClient()
+    protected function getClient(): ClientInterface
     {
         if (null !== $this->client) {
             return $this->client;
@@ -246,7 +222,7 @@ class AuthorizeController
 
         if (null === $clientId = $request->get('client_id')) {
             $formData = $request->get($this->authorizeForm->getName(), []);
-            $clientId = isset($formData['client_id']) ? $formData['client_id'] : null;
+            $clientId = $formData['client_id'] ?? null;
         }
 
         $this->client = $this->clientManager->findClientByPublicId($clientId);
@@ -258,25 +234,11 @@ class AuthorizeController
         return $this->client;
     }
 
-    /**
-     * @throws \RuntimeException
-     */
-    protected function renderAuthorize(array $data, EngineInterface $engine, string $engineType): Response
-    {
-        return $engine->renderResponse(
-            '@FOSOAuthServer/Authorize/authorize.html.'.$engineType,
-            $data
-        );
-    }
-
-    /**
-     * @return null|Request
-     */
-    private function getCurrentRequest()
+    private function getCurrentRequest(): ?Request
     {
         $request = $this->requestStack->getCurrentRequest();
         if (null === $request) {
-            throw new \RuntimeException('No current request.');
+            throw new RuntimeException('No current request.');
         }
 
         return $request;
