@@ -22,6 +22,7 @@ use Symfony\Component\Console\Style\SymfonyStyle;
 
 class CreateClientCommand extends Command
 {
+    /** @var ClientManagerInterface */
     private $clientManager;
 
     public function __construct(ClientManagerInterface $clientManager)
@@ -34,7 +35,7 @@ class CreateClientCommand extends Command
     /**
      * {@inheritdoc}
      */
-    protected function configure()
+    protected function configure(): void
     {
         parent::configure();
 
