@@ -13,6 +13,7 @@ declare(strict_types=1);
 
 namespace FOS\OAuthServerBundle\DependencyInjection;
 
+use InvalidArgumentException;
 use Symfony\Component\Config\Definition\Builder\ArrayNodeDefinition;
 use Symfony\Component\Config\Definition\Builder\TreeBuilder;
 use Symfony\Component\Config\Definition\ConfigurationInterface;
@@ -20,19 +21,20 @@ use Symfony\Component\Config\Definition\ConfigurationInterface;
 /**
  * This is the class that validates and merges configuration from your app/config files.
  *
- * To learn more see {@link http://symfony.com/doc/current/cookbook/bundles/extension.html#cookbook-bundles-extension-config-class}
+ * To learn more see
+ * {@link http://symfony.com/doc/current/cookbook/bundles/extension.html#cookbook-bundles-extension-config-class}
  */
 class Configuration implements ConfigurationInterface
 {
     /**
      * {@inheritdoc}
      */
-    public function getConfigTreeBuilder()
+    public function getConfigTreeBuilder(): TreeBuilder
     {
-        $treeBuilder = new TreeBuilder();
+        $treeBuilder = new TreeBuilder('fos_oauth_server');
 
         /** @var ArrayNodeDefinition $rootNode */
-        $rootNode = $treeBuilder->root('fos_oauth_server');
+        $rootNode = $treeBuilder->getRootNode();
 
         $supportedDrivers = ['orm', 'mongodb', 'propel', 'custom'];
 
@@ -43,20 +45,31 @@ class Configuration implements ConfigurationInterface
                         return $v;
                     }
 
-                    if (empty($v['service']['client_manager']) || $v['service']['client_manager'] === 'fos_oauth_server.client_manager.default') {
-                        throw new \InvalidArgumentException('The service client_manager must be set explicitly for custom db_driver.');
+                    if (empty($v['service']['client_manager'])
+                        ||
+                        $v['service']['client_manager'] === 'fos_oauth_server.client_manager.default'
+                    ) {
+                        throw new InvalidArgumentException('The service client_manager must be set explicitly for custom db_driver.');
                     }
 
-                    if (empty($v['service']['access_token_manager']) || $v['service']['access_token_manager'] === 'fos_oauth_server.access_token_manager.default') {
-                        throw new \InvalidArgumentException('The service access_token_manager must be set explicitly for custom db_driver.');
+                    if (empty($v['service']['access_token_manager'])
+                        ||
+                        $v['service']['access_token_manager'] === 'fos_oauth_server.access_token_manager.default') {
+                        throw new InvalidArgumentException('The service access_token_manager must be set explicitly for custom db_driver.');
                     }
 
-                    if (empty($v['service']['refresh_token_manager']) || $v['service']['refresh_token_manager'] === 'fos_oauth_server.refresh_token_manager.default') {
-                        throw new \InvalidArgumentException('The service refresh_token_manager must be set explicitly for custom db_driver.');
+                    if (empty($v['service']['refresh_token_manager'])
+                        ||
+                        $v['service']['refresh_token_manager'] === 'fos_oauth_server.refresh_token_manager.default'
+                    ) {
+                        throw new InvalidArgumentException('The service refresh_token_manager must be set explicitly for custom db_driver.');
                     }
 
-                    if (empty($v['service']['auth_code_manager']) || $v['service']['auth_code_manager'] === 'fos_oauth_server.auth_code_manager.default') {
-                        throw new \InvalidArgumentException('The service auth_code_manager must be set explicitly for custom db_driver.');
+                    if (empty($v['service']['auth_code_manager'])
+                        ||
+                        $v['service']['auth_code_manager'] === 'fos_oauth_server.auth_code_manager.default'
+                    ) {
+                        throw new InvalidArgumentException('The service auth_code_manager must be set explicitly for custom db_driver.');
                     }
 
                     return $v;
@@ -66,7 +79,10 @@ class Configuration implements ConfigurationInterface
                 ->scalarNode('db_driver')
                     ->validate()
                         ->ifNotInArray($supportedDrivers)
-                        ->thenInvalid('The driver %s is not supported. Please choose one of '.json_encode($supportedDrivers))
+                        ->thenInvalid(
+                            'The driver %s is not supported. Please choose one of '
+                            .json_encode($supportedDrivers)
+                        )
                     ->end()
                     ->isRequired()
                     ->cannotBeEmpty()
@@ -81,12 +97,11 @@ class Configuration implements ConfigurationInterface
 
         $this->addAuthorizeSection($rootNode);
         $this->addServiceSection($rootNode);
-        $this->addTemplateSection($rootNode);
 
         return $treeBuilder;
     }
 
-    private function addAuthorizeSection(ArrayNodeDefinition $node)
+    private function addAuthorizeSection(ArrayNodeDefinition $node): void
     {
         $node
             ->children()
@@ -97,9 +112,12 @@ class Configuration implements ConfigurationInterface
                         ->arrayNode('form')
                             ->addDefaultsIfNotSet()
                             ->children()
-                                ->scalarNode('type')->defaultValue('fos_oauth_server_authorize')->end()
-                                ->scalarNode('handler')->defaultValue('fos_oauth_server.authorize.form.handler.default')->end()
-                                ->scalarNode('name')->defaultValue('fos_oauth_server_authorize_form')->cannotBeEmpty()->end()
+                                ->scalarNode('type')
+                                    ->defaultValue('fos_oauth_server_authorize')->end()
+                                ->scalarNode('handler')
+                                    ->defaultValue('fos_oauth_server.authorize.form.handler.default')->end()
+                                ->scalarNode('name')
+                                    ->defaultValue('fos_oauth_server_authorize_form')->cannotBeEmpty()->end()
                                 ->arrayNode('validation_groups')
                                     ->prototype('scalar')->end()
                                     ->defaultValue(['Authorize', 'Default'])
@@ -112,7 +130,7 @@ class Configuration implements ConfigurationInterface
         ;
     }
 
-    private function addServiceSection(ArrayNodeDefinition $node)
+    private function addServiceSection(ArrayNodeDefinition $node): void
     {
         $node
             ->addDefaultsIfNotSet()
@@ -120,32 +138,24 @@ class Configuration implements ConfigurationInterface
                 ->arrayNode('service')
                     ->addDefaultsIfNotSet()
                         ->children()
-                            ->scalarNode('storage')->defaultValue('fos_oauth_server.storage.default')->cannotBeEmpty()->end()
-                            ->scalarNode('user_provider')->defaultNull()->end()
-                            ->scalarNode('client_manager')->defaultValue('fos_oauth_server.client_manager.default')->end()
-                            ->scalarNode('access_token_manager')->defaultValue('fos_oauth_server.access_token_manager.default')->end()
-                            ->scalarNode('refresh_token_manager')->defaultValue('fos_oauth_server.refresh_token_manager.default')->end()
-                            ->scalarNode('auth_code_manager')->defaultValue('fos_oauth_server.auth_code_manager.default')->end()
+                            ->scalarNode('storage')
+                                ->defaultValue('fos_oauth_server.storage.default')->cannotBeEmpty()->end()
+                            ->scalarNode('user_provider')
+                                ->defaultNull()->end()
+                            ->scalarNode('client_manager')
+                                ->defaultValue('fos_oauth_server.client_manager.default')->end()
+                            ->scalarNode('access_token_manager')
+                                ->defaultValue('fos_oauth_server.access_token_manager.default')->end()
+                            ->scalarNode('refresh_token_manager')
+                                ->defaultValue('fos_oauth_server.refresh_token_manager.default')->end()
+                            ->scalarNode('auth_code_manager')
+                                ->defaultValue('fos_oauth_server.auth_code_manager.default')->end()
                             ->arrayNode('options')
                                 ->useAttributeAsKey('key')
                                 ->treatNullLike([])
                                 ->prototype('variable')->end()
                             ->end()
                         ->end()
-                    ->end()
-                ->end()
-            ->end()
-        ;
-    }
-
-    private function addTemplateSection(ArrayNodeDefinition $node)
-    {
-        $node
-            ->children()
-                ->arrayNode('template')
-                    ->addDefaultsIfNotSet()
-                    ->children()
-                        ->scalarNode('engine')->defaultValue('twig')->end()
                     ->end()
                 ->end()
             ->end()
