@@ -30,8 +30,7 @@ class AuthCodeManager extends BaseAuthCodeManager
     protected $class;
 
     /**
-     * @param EntityManagerInterface $em
-     * @param string                 $class
+     * @param string $class
      */
     public function __construct(EntityManagerInterface $em, $class)
     {
