@@ -60,7 +60,7 @@ class OAuthProvider implements AuthenticationProviderInterface
     /**
      * @param OAuthToken&TokenInterface $token
      *
-     * @return null|OAuthToken
+     * @return OAuthToken|null
      */
     public function authenticate(TokenInterface $token)
     {
@@ -85,12 +85,7 @@ class OAuthProvider implements AuthenticationProviderInterface
                 try {
                     $this->userChecker->checkPreAuth($user);
                 } catch (AccountStatusException $e) {
-                    throw new OAuth2AuthenticateException(Response::HTTP_UNAUTHORIZED,
-                            OAuth2::TOKEN_TYPE_BEARER,
-                            $this->serverService->getVariable(OAuth2::CONFIG_WWW_REALM),
-                            'access_denied',
-                            $e->getMessage()
-                        );
+                    throw new OAuth2AuthenticateException(Response::HTTP_UNAUTHORIZED, OAuth2::TOKEN_TYPE_BEARER, $this->serverService->getVariable(OAuth2::CONFIG_WWW_REALM), 'access_denied', $e->getMessage());
                 }
 
                 $token->setUser($user);
@@ -114,12 +109,7 @@ class OAuthProvider implements AuthenticationProviderInterface
                 try {
                     $this->userChecker->checkPostAuth($user);
                 } catch (AccountStatusException $e) {
-                    throw new OAuth2AuthenticateException(Response::HTTP_UNAUTHORIZED,
-                            OAuth2::TOKEN_TYPE_BEARER,
-                            $this->serverService->getVariable(OAuth2::CONFIG_WWW_REALM),
-                            'access_denied',
-                            $e->getMessage()
-                        );
+                    throw new OAuth2AuthenticateException(Response::HTTP_UNAUTHORIZED, OAuth2::TOKEN_TYPE_BEARER, $this->serverService->getVariable(OAuth2::CONFIG_WWW_REALM), 'access_denied', $e->getMessage());
                 }
 
                 $token->setUser($user);
