@@ -16,19 +16,18 @@ namespace FOS\OAuthServerBundle\Security\Firewall;
 use FOS\OAuthServerBundle\Security\Authentication\Token\OAuthToken;
 use OAuth2\OAuth2;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\HttpKernel\Event\GetResponseEvent;
+use Symfony\Component\HttpKernel\Event\RequestEvent;
 use Symfony\Component\Security\Core\Authentication\AuthenticationManagerInterface;
 use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
 use Symfony\Component\Security\Core\Exception\AuthenticationException;
-use Symfony\Component\Security\Http\Firewall\ListenerInterface;
 
 /**
  * OAuthListener class.
  *
  * @author Arnaud Le Blanc <arnaud.lb@gmail.com>
  */
-class OAuthListener implements ListenerInterface
+class OAuthListener
 {
     /**
      * @var TokenStorageInterface
@@ -48,19 +47,23 @@ class OAuthListener implements ListenerInterface
     /**
      * @param TokenStorageInterface          $tokenStorage          the token storage
      * @param AuthenticationManagerInterface $authenticationManager the authentication manager
-     * @param OAuth2                         $serverService
      */
-    public function __construct(TokenStorageInterface $tokenStorage, AuthenticationManagerInterface $authenticationManager, OAuth2 $serverService)
-    {
+    public function __construct(
+        TokenStorageInterface $tokenStorage,
+        AuthenticationManagerInterface $authenticationManager,
+        OAuth2 $serverService
+    ) {
         $this->tokenStorage = $tokenStorage;
         $this->authenticationManager = $authenticationManager;
         $this->serverService = $serverService;
     }
 
-    /**
-     * @param GetResponseEvent $event the event
-     */
-    public function handle(GetResponseEvent $event)
+    public function __invoke(RequestEvent $event)
+    {
+        $this->handle($event);
+    }
+
+    public function handle(RequestEvent $event): void
     {
         if (null === $oauthToken = $this->serverService->getBearerToken($event->getRequest(), true)) {
             return;
@@ -70,14 +73,12 @@ class OAuthListener implements ListenerInterface
         $token->setToken($oauthToken);
 
         try {
-            $returnValue = $this->authenticationManager->authenticate($token);
+            $authenticateResult = $this->authenticationManager->authenticate($token);
 
-            if ($returnValue instanceof TokenInterface) {
-                return $this->tokenStorage->setToken($returnValue);
-            }
-
-            if ($returnValue instanceof Response) {
-                return $event->setResponse($returnValue);
+            if ($authenticateResult instanceof TokenInterface) {
+                $this->tokenStorage->setToken($authenticateResult);
+            } elseif ($authenticateResult instanceof Response) {
+                $event->setResponse($authenticateResult);
             }
         } catch (AuthenticationException $e) {
             if (null !== $p = $e->getPrevious()) {
