@@ -14,6 +14,8 @@ declare(strict_types=1);
 namespace FOS\OAuthServerBundle\Tests\DependencyInjection\Compiler;
 
 use FOS\OAuthServerBundle\DependencyInjection\Compiler\RequestStackCompilerPass;
+use PHPUnit\Framework\MockObject\MockObject;
+use PHPUnit\Framework\TestCase;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Definition;
 use Symfony\Component\DependencyInjection\Reference;
@@ -23,7 +25,7 @@ use Symfony\Component\DependencyInjection\Reference;
  *
  * @author Nikola Petkanski <nikola@petkanski.com>
  */
-class RequestStackCompilerPassTest extends \PHPUnit\Framework\TestCase
+class RequestStackCompilerPassTest extends TestCase
 {
     /**
      * @var RequestStackCompilerPass
@@ -31,11 +33,11 @@ class RequestStackCompilerPassTest extends \PHPUnit\Framework\TestCase
     protected $instance;
 
     /**
-     * @var \PHPUnit_Framework_MockObject_MockObject|ContainerBuilder
+     * @var MockObject|ContainerBuilder
      */
     protected $container;
 
-    public function setUp()
+    public function setUp(): void
     {
         $this->container = $this->getMockBuilder(ContainerBuilder::class)
             ->disableOriginalConstructor()
@@ -51,7 +53,7 @@ class RequestStackCompilerPassTest extends \PHPUnit\Framework\TestCase
         parent::setUp();
     }
 
-    public function testProcessWithoutRequestStackDoesNothing()
+    public function testProcessWithoutRequestStackDoesNothing(): void
     {
         $this->container
             ->expects($this->once())
@@ -60,10 +62,10 @@ class RequestStackCompilerPassTest extends \PHPUnit\Framework\TestCase
             ->willReturn(true)
         ;
 
-        $this->assertNull($this->instance->process($this->container));
+        self::assertNull($this->instance->process($this->container));
     }
 
-    public function testProcess()
+    public function testProcess(): void
     {
         $this->container
             ->expects($this->once())
@@ -96,6 +98,6 @@ class RequestStackCompilerPassTest extends \PHPUnit\Framework\TestCase
             ->willReturn(null)
         ;
 
-        $this->assertNull($this->instance->process($this->container));
+        self::assertNull($this->instance->process($this->container));
     }
 }
