@@ -13,12 +13,11 @@ declare(strict_types=1);
 
 namespace FOS\OAuthServerBundle\Tests\Document;
 
-use Doctrine\MongoDB\Query\Builder;
 use Doctrine\ODM\MongoDB\DocumentManager;
-use Doctrine\ODM\MongoDB\DocumentRepository;
-use Doctrine\ORM\AbstractQuery;
+use Doctrine\ODM\MongoDB\Repository\DocumentRepository;
 use FOS\OAuthServerBundle\Document\AuthCodeManager;
-use FOS\OAuthServerBundle\Model\AuthCodeInterface;
+use FOS\OAuthServerBundle\Tests\TestCase;
+use PHPUnit\Framework\MockObject\MockObject;
 
 /**
  * @group time-sensitive
@@ -27,15 +26,15 @@ use FOS\OAuthServerBundle\Model\AuthCodeInterface;
  *
  * @author Nikola Petkanski <nikola@petkanski.com>
  */
-class AuthCodeManagerTest extends \PHPUnit\Framework\TestCase
+class AuthCodeManagerTest extends TestCase
 {
     /**
-     * @var \PHPUnit_Framework_MockObject_MockObject|DocumentManager
+     * @var MockObject|DocumentManager
      */
     protected $documentManager;
 
     /**
-     * @var \PHPUnit_Framework_MockObject_MockObject|DocumentRepository
+     * @var MockObject|DocumentRepository
      */
     protected $repository;
 
@@ -49,9 +48,9 @@ class AuthCodeManagerTest extends \PHPUnit\Framework\TestCase
      */
     protected $instance;
 
-    public function setUp()
+    public function setUp(): void
     {
-        if (!class_exists('\Doctrine\ODM\MongoDB\DocumentManager')) {
+        if (!class_exists(DocumentManager::class)) {
             $this->markTestSkipped('Doctrine MongoDB ODM has to be installed for this test to run.');
         }
 
@@ -63,7 +62,7 @@ class AuthCodeManagerTest extends \PHPUnit\Framework\TestCase
             ->disableOriginalConstructor()
             ->getMock()
         ;
-        $this->className = 'TestClassName'.\random_bytes(5);
+        $this->className = 'TestClassName';
 
         $this->documentManager
             ->expects($this->once())
@@ -77,142 +76,14 @@ class AuthCodeManagerTest extends \PHPUnit\Framework\TestCase
         parent::setUp();
     }
 
-    public function testConstructWillSetParameters()
+    public function testConstructWillSetParameters(): void
     {
-        $this->assertAttributeSame($this->documentManager, 'dm', $this->instance);
-        $this->assertAttributeSame($this->className, 'class', $this->instance);
+        self::assertObjectPropertySame($this->documentManager, $this->instance, 'dm');
+        self::assertSame($this->className, $this->instance->getClass());
     }
 
-    public function testGetClassWillReturnClassName()
+    public function testGetClassWillReturnClassName(): void
     {
-        $this->assertSame($this->className, $this->instance->getClass());
-    }
-
-    public function testFindAuthCodeBy()
-    {
-        $randomResult = \random_bytes(10);
-        $criteria = [
-            \random_bytes(10),
-        ];
-
-        $this->repository
-            ->expects($this->once())
-            ->method('findOneBy')
-            ->with($criteria)
-            ->willReturn($randomResult)
-        ;
-
-        $this->assertSame($randomResult, $this->instance->findAuthCodeBy($criteria));
-    }
-
-    public function testUpdateAuthCode()
-    {
-        $authCode = $this->getMockBuilder(AuthCodeInterface::class)
-            ->disableOriginalConstructor()
-            ->getMock()
-        ;
-
-        $this->documentManager
-            ->expects($this->once())
-            ->method('persist')
-            ->with($authCode)
-            ->willReturn(null)
-        ;
-
-        $this->documentManager
-            ->expects($this->once())
-            ->method('flush')
-            ->with()
-            ->willReturn(null)
-        ;
-
-        $this->assertNull($this->instance->updateAuthCode($authCode));
-    }
-
-    public function testDeleteAuthCode()
-    {
-        $authCode = $this->getMockBuilder(AuthCodeInterface::class)
-            ->disableOriginalConstructor()
-            ->getMock()
-        ;
-
-        $this->documentManager
-            ->expects($this->once())
-            ->method('remove')
-            ->with($authCode)
-            ->willReturn(null)
-        ;
-
-        $this->documentManager
-            ->expects($this->once())
-            ->method('flush')
-            ->with()
-            ->willReturn(null)
-        ;
-
-        $this->assertNull($this->instance->deleteAuthCode($authCode));
-    }
-
-    public function testDeleteExpired()
-    {
-        $queryBuilder = $this->getMockBuilder(Builder::class)
-            ->disableOriginalConstructor()
-            ->getMock()
-        ;
-
-        $this->repository
-            ->expects($this->once())
-            ->method('createQueryBuilder')
-            ->with()
-            ->willReturn($queryBuilder)
-        ;
-
-        $queryBuilder
-            ->expects($this->once())
-            ->method('remove')
-            ->with()
-            ->willReturn($queryBuilder)
-        ;
-
-        $queryBuilder
-            ->expects($this->once())
-            ->method('field')
-            ->with('expiresAt')
-            ->willReturn($queryBuilder)
-        ;
-
-        $queryBuilder
-            ->expects($this->once())
-            ->method('lt')
-            ->with(time())
-            ->willReturn($queryBuilder)
-        ;
-
-        $query = $this->getMockBuilder(AbstractQuery::class)
-            ->disableOriginalConstructor()
-            ->getMock()
-        ;
-
-        $queryBuilder
-            ->expects($this->once())
-            ->method('getQuery')
-            ->with([
-                'safe' => true,
-            ])
-            ->willReturn($query)
-        ;
-
-        $data = [
-            'n' => \random_bytes(10),
-        ];
-
-        $query
-            ->expects($this->once())
-            ->method('execute')
-            ->with()
-            ->willReturn($data)
-        ;
-
-        $this->assertSame($data['n'], $this->instance->deleteExpired());
+        self::assertSame($this->className, $this->instance->getClass());
     }
 }
