@@ -13,12 +13,12 @@ declare(strict_types=1);
 
 namespace FOS\OAuthServerBundle\Tests\Document;
 
-use Doctrine\MongoDB\Query\Query;
 use Doctrine\ODM\MongoDB\DocumentManager;
-use Doctrine\ODM\MongoDB\DocumentRepository;
-use Doctrine\ODM\MongoDB\Query\Builder;
+use Doctrine\ODM\MongoDB\Repository\DocumentRepository;
 use FOS\OAuthServerBundle\Document\AccessToken;
 use FOS\OAuthServerBundle\Document\TokenManager;
+use PHPUnit\Framework\MockObject\MockObject;
+use PHPUnit\Framework\TestCase;
 
 /**
  * @group time-sensitive
@@ -27,7 +27,7 @@ use FOS\OAuthServerBundle\Document\TokenManager;
  *
  * @author Nikola Petkanski <nikola@petkanski.com>
  */
-class TokenManagerTest extends \PHPUnit\Framework\TestCase
+class TokenManagerTest extends TestCase
 {
     /**
      * @var string
@@ -35,12 +35,12 @@ class TokenManagerTest extends \PHPUnit\Framework\TestCase
     protected $className;
 
     /**
-     * @var \PHPUnit_Framework_MockObject_MockObject|DocumentManager
+     * @var MockObject|DocumentManager
      */
     protected $documentManager;
 
     /**
-     * @var \PHPUnit_Framework_MockObject_MockObject|DocumentRepository
+     * @var MockObject|DocumentRepository
      */
     protected $repository;
 
@@ -49,9 +49,9 @@ class TokenManagerTest extends \PHPUnit\Framework\TestCase
      */
     protected $instance;
 
-    public function setUp()
+    public function setUp(): void
     {
-        if (!class_exists('\Doctrine\ODM\MongoDB\DocumentManager')) {
+        if (!class_exists(DocumentManager::class)) {
             $this->markTestSkipped('Doctrine MongoDB ODM has to be installed for this test to run.');
         }
 
@@ -75,134 +75,8 @@ class TokenManagerTest extends \PHPUnit\Framework\TestCase
         $this->instance = new TokenManager($this->documentManager, $this->className);
     }
 
-    public function testFindTokenByToken()
+    public function testGetClass(): void
     {
-        $randomToken = \random_bytes(5);
-        $randomResult = \random_bytes(5);
-
-        $this->repository
-            ->expects($this->once())
-            ->method('findOneBy')
-            ->with([
-                'token' => $randomToken,
-            ])
-            ->willReturn($randomResult)
-        ;
-
-        $this->assertSame($randomResult, $this->instance->findTokenByToken($randomToken));
-    }
-
-    public function testUpdateTokenPersistsAndFlushes()
-    {
-        $token = $this->getMockBuilder(AccessToken::class)
-            ->disableOriginalConstructor()
-            ->getMock()
-        ;
-
-        $this->documentManager
-            ->expects($this->once())
-            ->method('persist')
-            ->with($token)
-        ;
-
-        $this->documentManager
-            ->expects($this->once())
-            ->method('flush')
-            ->with()
-        ;
-
-        $this->assertNull($this->instance->updateToken($token));
-    }
-
-    public function testGetClass()
-    {
-        $this->assertSame($this->className, $this->instance->getClass());
-    }
-
-    public function testDeleteToken()
-    {
-        $token = $this->getMockBuilder(AccessToken::class)
-            ->disableOriginalConstructor()
-            ->getMock()
-        ;
-
-        $this->documentManager
-            ->expects($this->once())
-            ->method('remove')
-            ->with($token)
-            ->willReturn(null)
-        ;
-
-        $this->documentManager
-            ->expects($this->once())
-            ->method('flush')
-            ->with()
-            ->willReturn(null)
-        ;
-
-        $this->assertNull($this->instance->deleteToken($token));
-    }
-
-    public function testDeleteExpired()
-    {
-        $queryBuilder = $this->getMockBuilder(Builder::class)
-            ->disableOriginalConstructor()
-            ->getMock()
-        ;
-
-        $this->repository
-            ->expects($this->once())
-            ->method('createQueryBuilder')
-            ->with()
-            ->willReturn($queryBuilder)
-        ;
-
-        $queryBuilder
-            ->expects($this->once())
-            ->method('remove')
-            ->with()
-            ->willReturn($queryBuilder)
-        ;
-
-        $queryBuilder
-            ->expects($this->once())
-            ->method('field')
-            ->with('expiresAt')
-            ->willReturn($queryBuilder)
-        ;
-
-        $queryBuilder
-            ->expects($this->once())
-            ->method('lt')
-            ->with(time())
-            ->willReturn($queryBuilder)
-        ;
-
-        $query = $this->getMockBuilder(Query::class)
-            ->disableOriginalConstructor()
-            ->getMock()
-        ;
-
-        $queryBuilder
-            ->expects($this->once())
-            ->method('getQuery')
-            ->with([
-                'safe' => true,
-            ])
-            ->willReturn($query)
-        ;
-
-        $data = [
-            'n' => \random_bytes(5),
-        ];
-
-        $query
-            ->expects($this->once())
-            ->method('execute')
-            ->with()
-            ->willReturn($data)
-        ;
-
-        $this->assertSame($data['n'], $this->instance->deleteExpired());
+        self::assertSame($this->className, $this->instance->getClass());
     }
 }
