@@ -15,28 +15,35 @@ namespace FOS\OAuthServerBundle\Tests\Form\Handler;
 
 use FOS\OAuthServerBundle\Form\Handler\AuthorizeFormHandler;
 use FOS\OAuthServerBundle\Form\Model\Authorize;
-use Symfony\Bundle\FrameworkBundle\Tests\Fixtures\Serialization\Author;
+use FOS\OAuthServerBundle\Tests\TestCase;
+use InvalidArgumentException;
+use PHPUnit\Framework\MockObject\MockObject;
+use function random_bytes;
+use ReflectionException;
+use ReflectionMethod;
+use ReflectionObject;
+use stdClass;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\Form\FormInterface;
 use Symfony\Component\HttpFoundation\ParameterBag;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\RequestStack;
 
-/**
- * Class AuthorizeFormHandlerTest.
- *
- * @author Nikola Petkanski <nikola@petkanski.com>
- */
-class AuthorizeFormHandlerTest extends \PHPUnit\Framework\TestCase
+class AuthorizeFormHandlerTest extends TestCase
 {
+    /** @var FormInterface | MockObject */
     protected $form;
 
+    /** @var Request | MockObject */
     protected $request;
 
+    /** @var ParameterBag | MockObject */
     protected $requestQuery;
 
+    /** @var ParameterBag | MockObject */
     protected $requestRequest;
 
+    /** @var ContainerInterface | MockObject */
     protected $container;
 
     /**
@@ -44,7 +51,7 @@ class AuthorizeFormHandlerTest extends \PHPUnit\Framework\TestCase
      */
     protected $instance;
 
-    public function setUp()
+    public function setUp(): void
     {
         $this->form = $this->getMockBuilder(FormInterface::class)
             ->disableOriginalConstructor()
@@ -78,7 +85,7 @@ class AuthorizeFormHandlerTest extends \PHPUnit\Framework\TestCase
         parent::setUp();
     }
 
-    public function testConstructWillAcceptRequestObjectAsRequest()
+    public function testConstructWillAcceptRequestObjectAsRequest(): void
     {
         $request = $this->getMockBuilder(Request::class)
             ->disableOriginalConstructor()
@@ -87,11 +94,10 @@ class AuthorizeFormHandlerTest extends \PHPUnit\Framework\TestCase
 
         $this->instance = new AuthorizeFormHandler($this->form, $request);
 
-        $this->assertAttributeSame($this->form, 'form', $this->instance);
-        $this->assertAttributeSame($request, 'requestStack', $this->instance);
+        $this->assertAttributesWereSet($request);
     }
 
-    public function testConstructWillAcceptRequestStackObjectAsRequest()
+    public function testConstructWillAcceptRequestStackObjectAsRequest(): void
     {
         $requestStack = $this->getMockBuilder(RequestStack::class)
             ->disableOriginalConstructor()
@@ -100,38 +106,35 @@ class AuthorizeFormHandlerTest extends \PHPUnit\Framework\TestCase
 
         $this->instance = new AuthorizeFormHandler($this->form, $requestStack);
 
-        $this->assertAttributeSame($this->form, 'form', $this->instance);
-        $this->assertAttributeSame($requestStack, 'requestStack', $this->instance);
+        $this->assertAttributesWereSet($requestStack);
     }
 
     public function testConstructWillAcceptNullAsRequest()
     {
         $this->instance = new AuthorizeFormHandler($this->form, null);
-        $this->assertAttributeSame($this->form, 'form', $this->instance);
-        $this->assertAttributeSame(null, 'requestStack', $this->instance);
+        $this->assertAttributesWereSet(null);
 
         $this->instance = new AuthorizeFormHandler($this->form);
-        $this->assertAttributeSame($this->form, 'form', $this->instance);
-        $this->assertAttributeSame(null, 'requestStack', $this->instance);
+        $this->assertAttributesWereSet(null);
     }
 
-    public function testConstructWillThrowException()
+    public function testConstructWillThrowException(): void
     {
         $exceptionMessage = sprintf(
             'Argument 2 of %s must be an instanceof RequestStack or Request',
             AuthorizeFormHandler::class
         );
 
-        $this->expectException(\InvalidArgumentException::class);
+        $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessage($exceptionMessage);
 
-        new AuthorizeFormHandler($this->form, new \stdClass());
+        new AuthorizeFormHandler($this->form, new stdClass());
     }
 
-    public function testIsAcceptedWillProxyValueToFormData()
+    public function testIsAcceptedWillProxyValueToFormData(): void
     {
-        $data = new \stdClass();
-        $data->accepted = \random_bytes(10);
+        $data = new stdClass();
+        $data->accepted = random_bytes(10);
 
         $this->form
             ->expects($this->once())
@@ -140,15 +143,15 @@ class AuthorizeFormHandlerTest extends \PHPUnit\Framework\TestCase
             ->willReturn($data)
         ;
 
-        $this->assertSame($data->accepted, $this->instance->isAccepted());
+        self::assertSame($data->accepted, $this->instance->isAccepted());
     }
 
-    public function testIsRejectedWillNegateAcceptedValueFromFormData()
+    public function testIsRejectedWillNegateAcceptedValueFromFormData(): void
     {
-        $dataWithAcceptedValueFalse = new \stdClass();
+        $dataWithAcceptedValueFalse = new stdClass();
         $dataWithAcceptedValueFalse->accepted = false;
 
-        $dataWithAcceptedValueTrue = new \stdClass();
+        $dataWithAcceptedValueTrue = new stdClass();
         $dataWithAcceptedValueTrue->accepted = true;
 
         $this->form
@@ -160,14 +163,14 @@ class AuthorizeFormHandlerTest extends \PHPUnit\Framework\TestCase
             )
         ;
 
-        $this->assertTrue($this->instance->isRejected());
-        $this->assertFalse($this->instance->isRejected());
+        self::assertTrue($this->instance->isRejected());
+        self::assertFalse($this->instance->isRejected());
     }
 
-    public function testGetScopeWillProxyValueToFormData()
+    public function testGetScopeWillProxyValueToFormData(): void
     {
-        $data = new \stdClass();
-        $data->scope = \random_bytes(10);
+        $data = new stdClass();
+        $data->scope = random_bytes(10);
 
         $this->form
             ->expects($this->once())
@@ -176,16 +179,16 @@ class AuthorizeFormHandlerTest extends \PHPUnit\Framework\TestCase
             ->willReturn($data)
         ;
 
-        $this->assertSame($data->scope, $this->instance->getScope());
+        self::assertSame($data->scope, $this->instance->getScope());
     }
 
-    public function testGetCurrentRequestWillReturnRequestObject()
+    public function testGetCurrentRequestWillReturnRequestObject(): void
     {
         $method = $this->getReflectionMethod('getCurrentRequest');
-        $this->assertSame($this->request, $method->invoke($this->instance));
+        self::assertSame($this->request, $method->invoke($this->instance));
     }
 
-    public function testGetCurrentRequestWillReturnCurrentRequestFromRequestStack()
+    public function testGetCurrentRequestWillReturnCurrentRequestFromRequestStack(): void
     {
         $requestStack = $this->getMockBuilder(RequestStack::class)
             ->disableOriginalConstructor()
@@ -193,7 +196,7 @@ class AuthorizeFormHandlerTest extends \PHPUnit\Framework\TestCase
         ;
         $this->instance = new AuthorizeFormHandler($this->form, $requestStack);
 
-        $request = new \stdClass();
+        $request = new stdClass();
 
         $requestStack
             ->expects($this->once())
@@ -203,15 +206,15 @@ class AuthorizeFormHandlerTest extends \PHPUnit\Framework\TestCase
         ;
 
         $method = $this->getReflectionMethod('getCurrentRequest');
-        $this->assertSame($request, $method->invoke($this->instance));
+        self::assertSame($request, $method->invoke($this->instance));
     }
 
-    public function testGetCurrentRequestWillReturnRequestServiceFromContainerIfNoneIsSet()
+    public function testGetCurrentRequestWillReturnRequestServiceFromContainerIfNoneIsSet(): void
     {
         $this->instance = new AuthorizeFormHandler($this->form, null);
         $this->instance->setContainer($this->container);
 
-        $randomData = \random_bytes(10);
+        $randomData = random_bytes(10);
 
         $this->container
             ->expects($this->at(0))
@@ -221,22 +224,22 @@ class AuthorizeFormHandlerTest extends \PHPUnit\Framework\TestCase
         ;
 
         $method = $this->getReflectionMethod('getCurrentRequest');
-        $this->assertSame($randomData, $method->invoke($this->instance));
+        self::assertSame($randomData, $method->invoke($this->instance));
     }
 
     /**
      * @TODO Fix this behavior. This method MUST not modify $_GET.
      */
-    public function testOnSuccessWillReplaceGETSuperGlobal()
+    public function testOnSuccessWillReplaceGETSuperGlobal(): void
     {
         $method = $this->getReflectionMethod('onSuccess');
 
-        $data = new \stdClass();
-        $data->client_id = \random_bytes(10);
-        $data->response_type = \random_bytes(10);
-        $data->redirect_uri = \random_bytes(10);
-        $data->state = \random_bytes(10);
-        $data->scope = \random_bytes(10);
+        $data = new stdClass();
+        $data->client_id = random_bytes(10);
+        $data->response_type = random_bytes(10);
+        $data->redirect_uri = random_bytes(10);
+        $data->state = random_bytes(10);
+        $data->scope = random_bytes(10);
 
         $this->form
             ->expects($this->exactly(5))
@@ -255,12 +258,12 @@ class AuthorizeFormHandlerTest extends \PHPUnit\Framework\TestCase
             'scope' => $data->scope,
         ];
 
-        $this->assertNull($method->invoke($this->instance));
+        self::assertNull($method->invoke($this->instance));
 
-        $this->assertSame($expectedSuperGlobalValue, $_GET);
+        self::assertSame($expectedSuperGlobalValue, $_GET);
     }
 
-    public function testProcessWillReturnFalseIfRequestIsNull()
+    public function testProcessWillReturnFalseIfRequestIsNull(): void
     {
         $this->instance = new AuthorizeFormHandler($this->form, null);
         $this->instance->setContainer($this->container);
@@ -272,10 +275,10 @@ class AuthorizeFormHandlerTest extends \PHPUnit\Framework\TestCase
             ->willReturn(null)
         ;
 
-        $this->assertFalse($this->instance->process());
+        self::assertFalse($this->instance->process());
     }
 
-    public function testProcessWillSetFormData()
+    public function testProcessWillSetFormData(): void
     {
         $this->requestRequest
             ->expects($this->once())
@@ -285,8 +288,8 @@ class AuthorizeFormHandlerTest extends \PHPUnit\Framework\TestCase
         ;
 
         $dataMock = [
-            \random_bytes(10),
-            \random_bytes(10),
+            random_bytes(10),
+            random_bytes(10),
         ];
 
         $this->requestQuery
@@ -306,10 +309,10 @@ class AuthorizeFormHandlerTest extends \PHPUnit\Framework\TestCase
             ->willReturn($this->form)
         ;
 
-        $this->assertFalse($this->instance->process());
+        self::assertFalse($this->instance->process());
     }
 
-    public function testProcessWillHandleRequestOnPost()
+    public function testProcessWillHandleRequestOnPost(): void
     {
         $this->requestRequest
             ->expects($this->once())
@@ -319,8 +322,8 @@ class AuthorizeFormHandlerTest extends \PHPUnit\Framework\TestCase
         ;
 
         $dataMock = [
-            \random_bytes(10),
-            \random_bytes(10),
+            random_bytes(10),
+            random_bytes(10),
         ];
 
         $this->requestQuery
@@ -356,15 +359,22 @@ class AuthorizeFormHandlerTest extends \PHPUnit\Framework\TestCase
 
         $this->form
             ->expects($this->once())
+            ->method('isSubmitted')
+            ->with()
+            ->willReturn(true)
+        ;
+
+        $this->form
+            ->expects($this->once())
             ->method('isValid')
             ->with()
             ->willReturn(false)
         ;
 
-        $this->assertFalse($this->instance->process());
+        self::assertFalse($this->instance->process());
     }
 
-    public function testProcessWillHandleRequestOnPostAndWillProcessDataIfFormIsValid()
+    public function testProcessWillHandleRequestOnPostAndWillProcessDataIfFormIsValid(): void
     {
         $this->requestRequest
             ->expects($this->once())
@@ -373,12 +383,12 @@ class AuthorizeFormHandlerTest extends \PHPUnit\Framework\TestCase
             ->willReturn(true)
         ;
 
-        $query = new \stdClass();
-        $query->client_id = \random_bytes(10);
-        $query->response_type = \random_bytes(10);
-        $query->redirect_uri = \random_bytes(10);
-        $query->state = \random_bytes(10);
-        $query->scope = \random_bytes(10);
+        $query = new stdClass();
+        $query->client_id = random_bytes(10);
+        $query->response_type = random_bytes(10);
+        $query->redirect_uri = random_bytes(10);
+        $query->state = random_bytes(10);
+        $query->scope = random_bytes(10);
 
         $this->requestQuery
             ->expects($this->once())
@@ -415,6 +425,13 @@ class AuthorizeFormHandlerTest extends \PHPUnit\Framework\TestCase
 
         $this->form
             ->expects($this->once())
+            ->method('isSubmitted')
+            ->with()
+            ->willReturn(true)
+        ;
+
+        $this->form
+            ->expects($this->once())
             ->method('isValid')
             ->with()
             ->willReturn(true)
@@ -427,7 +444,7 @@ class AuthorizeFormHandlerTest extends \PHPUnit\Framework\TestCase
             ->willReturn($formData)
         ;
 
-        $this->assertSame([], $_GET);
+        self::assertSame([], $_GET);
 
         $expectedSuperGlobalValue = [
             'client_id' => $query->client_id,
@@ -437,22 +454,29 @@ class AuthorizeFormHandlerTest extends \PHPUnit\Framework\TestCase
             'scope' => $query->scope,
         ];
 
-        $this->assertTrue($this->instance->process());
+        self::assertTrue($this->instance->process());
 
-        $this->assertSame($expectedSuperGlobalValue, $_GET);
+        self::assertSame($expectedSuperGlobalValue, $_GET);
     }
 
     /**
-     * @param string $methodName
-     *
-     * @return \ReflectionMethod
+     * @throws ReflectionException
      */
-    protected function getReflectionMethod($methodName)
+    protected function getReflectionMethod(string $methodName): ReflectionMethod
     {
-        $reflectionObject = new \ReflectionObject($this->instance);
+        $reflectionObject = new ReflectionObject($this->instance);
         $reflectionMethod = $reflectionObject->getMethod($methodName);
         $reflectionMethod->setAccessible(true);
 
         return $reflectionMethod;
+    }
+
+    /**
+     * @param MockObject $request
+     */
+    private function assertAttributesWereSet(?MockObject $request): void
+    {
+        self::assertObjectPropertySame($this->form, $this->instance, 'form');
+        self::assertObjectPropertySame($request, $this->instance, 'requestStack');
     }
 }
