@@ -13,32 +13,75 @@ declare(strict_types=1);
 
 namespace FOS\OAuthServerBundle\Tests\Functional;
 
+use LogicException;
+use Symfony\Bundle\FrameworkBundle\KernelBrowser;
 use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
 use Symfony\Component\Filesystem\Filesystem;
-use Symfony\Component\HttpKernel\KernelInterface;
 
 abstract class TestCase extends WebTestCase
 {
     /**
-     * @var null|KernelInterface
+     * @var KernelBrowser
      */
-    protected static $kernel;
+    protected $client;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $fs = new Filesystem();
         $fs->remove(sys_get_temp_dir().'/FOSOAuthServerBundle/');
     }
 
-    protected function tearDown()
+    /**
+     * Client response assertion of status code and response content.
+     */
+    protected function assertResponse(int $statusCode, string $content, bool $fullFailOutput = false): void
     {
-        static::$kernel = null;
-    }
+        if (!($this->client instanceof KernelBrowser)) {
+            throw new LogicException('Test attempts to check response, but client does not exist; use createClient() to set the test case client property.');
+        }
 
-    protected static function createKernel(array $options = [])
-    {
-        $env = @$options['env'] ?: 'test';
+        $this->assertSame(
+            $statusCode,
+            $this->client->getResponse()->getStatusCode(),
+            sprintf('Failed asserting that response status code "%d" is "%d".', $this->client->getResponse()->getStatusCode(), $statusCode)
+        );
 
-        return new AppKernel($env, true);
+        $responseContent = $this->client->getResponse()->getContent();
+
+        if ('' === $responseContent && '' === $content) {
+            $this->assertTrue(true);
+
+            return;
+        }
+
+        if ('' === $responseContent) {
+            $this->fail(sprintf('Response content is empty, expected "%s".', $content));
+        } elseif ('' === $content) {
+            // this differs from assertStringContainsString, which does not
+            // fail on an empty string expectation
+            $this->fail($fullFailOutput || mb_strlen($responseContent) < 100
+                ? sprintf('Failed asserting that response "%s" is empty.', $responseContent)
+                : sprintf(
+                    'Failed asserting that response "%s ... %s" is empty.',
+                    mb_substr($responseContent, 0, 40),
+                    mb_substr($responseContent, mb_strlen($responseContent) - 40)
+                )
+            );
+        }
+
+        // not using assertStringContainsString to avoid full HTML doc in the
+        // fail message
+        if (mb_strpos($responseContent, $content) === false) {
+            $this->fail($fullFailOutput || mb_strlen($responseContent) < 100
+                ? sprintf('Failed asserting that response "%s" contains "%s".', $responseContent, $content)
+                : sprintf(
+                    'Failed asserting that response "%s ... %s" contains "%s".',
+                    mb_substr($responseContent, 0, 40),
+                    mb_substr($responseContent, mb_strlen($responseContent) - 40),
+                    $content
+                )
+            );
+        }
+        $this->assertTrue(true);
     }
 }
