@@ -19,11 +19,11 @@ use FOS\OAuthServerBundle\Propel\ClientQuery;
 
 class ClientManagerTest extends PropelTestCase
 {
-    const CLIENT_CLASS = 'FOS\OAuthServerBundle\Propel\Client';
+    const CLIENT_CLASS = Client::class;
 
     protected $manager;
 
-    public function setUp()
+    public function setUp(): void
     {
         parent::setUp();
 
@@ -33,17 +33,17 @@ class ClientManagerTest extends PropelTestCase
 
     public function testConstruct()
     {
-        $this->assertSame(self::CLIENT_CLASS, $this->manager->getClass());
+        self::assertSame(self::CLIENT_CLASS, $this->manager->getClass());
     }
 
     public function testCreateClass()
     {
-        $this->assertInstanceOf(self::CLIENT_CLASS, $this->manager->createClient());
+        self::assertInstanceOf(self::CLIENT_CLASS, $this->manager->createClient());
     }
 
     public function testUpdate()
     {
-        $client = $this->getMockBuilder('FOS\OAuthServerBundle\Propel\Client')
+        $client = $this->getMockBuilder(Client::class)
             ->disableOriginalConstructor()
             ->getMock()
         ;
@@ -57,7 +57,8 @@ class ClientManagerTest extends PropelTestCase
 
     public function testDelete()
     {
-        $client = $this->getMockBuilder('FOS\OAuthServerBundle\Propel\Client')
+        /** @var Client $client */
+        $client = $this->getMockBuilder(Client::class)
             ->disableOriginalConstructor()
             ->getMock()
         ;
@@ -73,19 +74,19 @@ class ClientManagerTest extends PropelTestCase
     {
         $client = $this->manager->findClientBy(['id' => '1', 'randomId' => '2345']);
 
-        $this->assertNull($client);
+        self::assertNull($client);
     }
 
     public function testFindClientWithInvalidCriteria()
     {
         $client = $this->manager->findClientBy(['randomId' => '2345']);
-        $this->assertNull($client);
+        self::assertNull($client);
 
         $client = $this->manager->findClientBy(['id' => '2345']);
-        $this->assertNull($client);
+        self::assertNull($client);
 
         $client = $this->manager->findClientBy(['foo' => '2345']);
-        $this->assertNull($client);
+        self::assertNull($client);
     }
 
     public function testFindClient()
@@ -93,8 +94,8 @@ class ClientManagerTest extends PropelTestCase
         $client = $this->createClient('2345');
         $return = $this->manager->findClientBy(['id' => '1', 'randomId' => '2345']);
 
-        $this->assertNotNull($return);
-        $this->assertSame($client, $return);
+        self::assertNotNull($return);
+        self::assertSame($client, $return);
     }
 
     public function testFindClientByPublicId()
@@ -102,28 +103,28 @@ class ClientManagerTest extends PropelTestCase
         $client = $this->createClient('12345');
         $return = $this->manager->findClientByPublicId('1_12345');
 
-        $this->assertNotNull($return);
-        $this->assertSame($client, $return);
+        self::assertNotNull($return);
+        self::assertSame($client, $return);
     }
 
     public function testFindClientByPublicIdReturnsNullIfNotFound()
     {
         $return = $this->manager->findClientByPublicId('1_12345');
 
-        $this->assertNull($return);
+        self::assertNull($return);
     }
 
     public function testFindClientByPublicIdReturnsNullIfInvalidPublicId()
     {
         $return = $this->manager->findClientByPublicId('1');
-        $this->assertNull($return);
+        self::assertNull($return);
 
         $return = $this->manager->findClientByPublicId('');
-        $this->assertNull($return);
+        self::assertNull($return);
 
         // invalid type
         // $return = $this->manager->findClientByPublicId(null);
-        // $this->assertNull($return);
+        // self::assertNull($return);
     }
 
     protected function createClient($randomId)
