@@ -15,17 +15,18 @@ namespace FOS\OAuthServerBundle\Tests\Util;
 
 use FOS\OAuthServerBundle\Util\Random;
 use phpmock\phpunit\PHPMock;
+use PHPUnit\Framework\TestCase;
 
 /**
  * Class RandomTest.
  *
  * @author Nikola Petkanski <nikola@petkanski.com
  */
-class RandomTest extends \PHPUnit\Framework\TestCase
+class RandomTest extends TestCase
 {
     use PHPMock;
 
-    public function setUp()
+    public function setUp(): void
     {
         parent::setUp();
     }
@@ -33,7 +34,7 @@ class RandomTest extends \PHPUnit\Framework\TestCase
     /**
      * @runInSeparateProcess
      */
-    public function testGenerateTokenWillUseRandomBytesIfAvailable()
+    public function testGenerateTokenWillUseRandomBytesIfAvailable(): void
     {
         $hashResult = \random_bytes(32);
 
@@ -57,6 +58,6 @@ class RandomTest extends \PHPUnit\Framework\TestCase
             ->willReturn($baseConvertResult)
         ;
 
-        $this->assertSame($baseConvertResult, Random::generateToken());
+        self::assertSame($baseConvertResult, Random::generateToken());
     }
 }
